HHA504_mysql_vm_vs_managed/requirements.txt
sqlalchemy>=1.4
pymysql
pandas
python-dotenv
